Move misplaced headers to canonical modules
