+++ gtsam/constrained/tests/simulated2DConstraints.h
@@ -21,9 +21,9 @@
 
 #include <gtsam/base/numericalDerivative.h>
 
-#include <gtsam/nonlinear/NonlinearEquality.h>
+#include <gtsam/constrained/NonlinearEquality.h>
 #include <gtsam/slam/BetweenFactor.h>
-#include <gtsam/slam/BoundingConstraint.h>
+#include <gtsam/constrained/BoundingConstraint.h>
 #include <tests/simulated2D.h>
 #include "gtsam/nonlinear/NonlinearFactor.h"
