+++ gtsam/inference/tests/testKruskal.cpp
@@ -18,7 +18,7 @@
 #include <CppUnitLite/TestHarness.h>
 #include <gtsam/base/MatrixConstants.h>
 #include <gtsam/base/TestableAssertions.h>
-#include <gtsam/base/kruskal.h>
+#include <gtsam/inference/kruskal.h>
 #include <gtsam/geometry/Rot3.h>
 #include <gtsam/inference/Symbol.h>
 #include <gtsam/linear/GaussianFactorGraph.h>
