+++ tests/testDoglegOptimizer.cpp
@@ -22,7 +22,7 @@
 #include <gtsam/geometry/Pose2.h>
 #include <gtsam/nonlinear/DoglegOptimizer.h>
 #include <gtsam/nonlinear/DoglegOptimizerImpl.h>
-#include <gtsam/nonlinear/NonlinearEquality.h>
+#include <gtsam/constrained/NonlinearEquality.h>
 #include <gtsam/slam/BetweenFactor.h>
 #include <gtsam/nonlinear/ISAM2.h>
 #include <gtsam/slam/SmartProjectionPoseFactor.h>
